Moved apportionment/mod.rs code to seat_assignment.rs and renamed all functions, structs, etc. accordingly. Added candidate_nomination.rs with structs.
