+++ backend/src/apportionment/api.rs
@@ -7,7 +7,7 @@ use utoipa::ToSchema;
 
 use crate::{
     APIError, ErrorResponse,
-    apportionment::{ApportionmentError, ApportionmentResult, apportionment},
+    apportionment::{ApportionmentError, SeatAssignmentResult, seat_assignment},
     authentication::Coordinator,
     data_entry::{
         repository::{PollingStationDataEntries, PollingStationResultsEntries},
@@ -21,7 +21,7 @@ use crate::{
 /// Election details response, including the election's candidate list (political groups) and its polling stations
 #[derive(Serialize, Deserialize, ToSchema, Debug)]
 pub struct ElectionApportionmentResponse {
-    pub apportionment: ApportionmentResult,
+    pub seat_assignment: SeatAssignmentResult,
     pub election_summary: ElectionSummary,
 }
 
@@ -59,9 +59,9 @@ pub async fn election_apportionment(
             .list_with_polling_stations(polling_stations_repo, election.id)
             .await?;
         let election_summary = ElectionSummary::from_results(&election, &results)?;
-        let apportionment = apportionment(election.number_of_seats.into(), &election_summary)?;
+        let seat_assignment = seat_assignment(election.number_of_seats.into(), &election_summary)?;
         Ok(Json(ElectionApportionmentResponse {
-            apportionment,
+            seat_assignment,
             election_summary,
         }))
     } else {
