+++ backend/tests/apportionment_integration_test.rs
@@ -44,10 +44,10 @@ async fn test_election_apportionment_works_for_less_than_19_seats(pool: SqlitePo
     // Ensure the response is what we expect
     assert_eq!(response.status(), StatusCode::OK);
     let body: ElectionApportionmentResponse = response.json().await.unwrap();
-    assert_eq!(body.apportionment.seats, 15);
-    assert_eq!(body.apportionment.quota, Fraction::new(204, 15));
-    assert_eq!(body.apportionment.steps.len(), 1);
-    let total_seats = get_total_seats_from_apportionment_result(body.apportionment);
+    assert_eq!(body.seat_assignment.seats, 15);
+    assert_eq!(body.seat_assignment.quota, Fraction::new(204, 15));
+    assert_eq!(body.seat_assignment.steps.len(), 1);
+    let total_seats = get_total_seats_from_apportionment_result(body.seat_assignment);
     assert_eq!(total_seats, vec![9, 6]);
 }
 
@@ -69,10 +69,10 @@ async fn test_election_apportionment_works_for_19_or_more_seats(pool: SqlitePool
     // Ensure the response is what we expect
     assert_eq!(response.status(), StatusCode::OK);
     let body: ElectionApportionmentResponse = response.json().await.unwrap();
-    assert_eq!(body.apportionment.seats, 29);
-    assert_eq!(body.apportionment.quota, Fraction::new(102, 29));
-    assert_eq!(body.apportionment.steps.len(), 1);
-    let total_seats = get_total_seats_from_apportionment_result(body.apportionment);
+    assert_eq!(body.seat_assignment.seats, 29);
+    assert_eq!(body.seat_assignment.quota, Fraction::new(102, 29));
+    assert_eq!(body.seat_assignment.steps.len(), 1);
+    let total_seats = get_total_seats_from_apportionment_result(body.seat_assignment);
     assert_eq!(total_seats, vec![17, 12]);
 }
