+++ frontend/app/component/apportionment/LargestAveragesFor19OrMoreSeatsTable.tsx
@@ -1,8 +1,8 @@
 import {
-  ApportionmentStep,
   LargestAverageAssignedSeat,
   PoliticalGroup,
   PoliticalGroupSeatAssignment,
+  SeatAssignmentStep,
 } from "@kiesraad/api";
 import { t } from "@kiesraad/i18n";
 import { Table } from "@kiesraad/ui";
@@ -11,7 +11,7 @@ import { cn } from "@kiesraad/util";
 import cls from "./Apportionment.module.css";
 
 interface LargestAveragesFor19OrMoreSeatsTableProps {
-  largestAverageSteps: ApportionmentStep[];
+  largestAverageSteps: SeatAssignmentStep[];
   finalStanding: PoliticalGroupSeatAssignment[];
   politicalGroups: PoliticalGroup[];
 }
@@ -30,7 +30,7 @@ export function LargestAveragesFor19OrMoreSeatsTable({
         <Table.Header>
           <Table.HeaderCell className={cn(cls.sticky, "text-align-r")}>{t("list")}</Table.HeaderCell>
           <Table.HeaderCell className={cls.sticky}>{t("list_name")}</Table.HeaderCell>
-          {largestAverageSteps.map((step: ApportionmentStep) => {
+          {largestAverageSteps.map((step: SeatAssignmentStep) => {
             return (
               <Table.HeaderCell key={step.residual_seat_number} className="text-align-r" span={2}>
                 {t("apportionment.residual_seat.singular")} {step.residual_seat_number}
@@ -51,7 +51,7 @@ export function LargestAveragesFor19OrMoreSeatsTable({
                 <Table.Cell className={cls.sticky}>
                   {politicalGroups[pg_seat_assignment.pg_number - 1]?.name || ""}
                 </Table.Cell>
-                {largestAverageSteps.map((step: ApportionmentStep) => {
+                {largestAverageSteps.map((step: SeatAssignmentStep) => {
                   const change = step.change as LargestAverageAssignedSeat;
                   const average = step.standing[pg_seat_assignment.pg_number - 1]?.next_votes_per_seat;
                   if (average) {
@@ -78,7 +78,7 @@ export function LargestAveragesFor19OrMoreSeatsTable({
             <Table.Cell className={cn(cls.sticky, "text-align-r", "nowrap", "bold")}>
               {t("apportionment.residual_seat_assigned_to_list")}
             </Table.Cell>
-            {largestAverageSteps.map((step: ApportionmentStep) => {
+            {largestAverageSteps.map((step: SeatAssignmentStep) => {
               const change = step.change as LargestAverageAssignedSeat;
               return (
                 <Table.NumberCell key={step.residual_seat_number} colSpan={2}>
