+++ frontend/app/component/apportionment/LargestAveragesForLessThan19SeatsTable.tsx
@@ -1,8 +1,8 @@
 import {
-  ApportionmentStep,
   LargestAverageAssignedSeat,
   PoliticalGroup,
   PoliticalGroupSeatAssignment,
+  SeatAssignmentStep,
 } from "@kiesraad/api";
 import { t } from "@kiesraad/i18n";
 import { Table } from "@kiesraad/ui";
@@ -11,7 +11,7 @@ import { cn } from "@kiesraad/util";
 import cls from "./Apportionment.module.css";
 
 interface LargestAveragesForLessThan19SeatsTableProps {
-  largestAverageSteps: ApportionmentStep[];
+  largestAverageSteps: SeatAssignmentStep[];
   finalStanding: PoliticalGroupSeatAssignment[];
   politicalGroups: PoliticalGroup[];
 }
