+++ frontend/app/component/apportionment/LargestRemaindersTable.tsx
@@ -1,8 +1,8 @@
 import {
-  ApportionmentStep,
   LargestRemainderAssignedSeat,
   PoliticalGroup,
   PoliticalGroupSeatAssignment,
+  SeatAssignmentStep,
 } from "@kiesraad/api";
 import { t } from "@kiesraad/i18n";
 import { Table } from "@kiesraad/ui";
@@ -11,7 +11,7 @@ import { cn } from "@kiesraad/util";
 import cls from "./Apportionment.module.css";
 
 interface LargestRemaindersTableProps {
-  largestRemainderSteps: ApportionmentStep[];
+  largestRemainderSteps: SeatAssignmentStep[];
   finalStanding: PoliticalGroupSeatAssignment[];
   politicalGroups: PoliticalGroup[];
 }
